requirements.txt
requests
bs4
csv
re
time
apache-airflow